Plant Disease Detection/user_interface/requirements.txt
numpy>=1.21.0
streamlit>=1.17.0
Pillow>=9.0.0
torch==2.5.0
torchvision>=0.14.0
matplotlib>=3.5.0
scikit-learn>=1.0.2
segmentation-models-pytorch
# albumentations